landrightgithubagent-main/python-agent/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
httpx>=0.27.0
pydantic>=2.0
python-dotenv>=1.0.0
supabase>=2.0.0
PyGithub>=2.4.0
anthropic>=0.39.0
apscheduler>=3.10.0
playwright>=1.40.0
groq>=0.4.0
